Add files via upload
Objective of the project¶
The code solves the group analysis of data taken with a "FLIR" thermograph. The objective of this code is a tool that processes hundreds of thermal images from an experiment conducted on birds of different ages, the images are treated in the following processes:
Import metadata from FLIR camera.
Convert the image to a raster image.
Sample or crop the raster image.
Cluster the sample image with Kmeans.
Statistical analysis of the clustered image.
Export the results in '.csv' format.